requirements.txt
websockets>=11.0
tk>=0.1.0
python-dotenv>=0.19.0 